requirements.txt
hikari>=2.0.0.dev122
alluka>=0.1.4
attrs>=23.1
